+++ homeassistant/helpers/config_validation.py
@@ -1126,26 +1126,6 @@ def key_value_schemas(
     return key_value_validator
 
 
-def get_absolute_description_key(domain: str, key: str) -> str:
-    """Return the absolute description key."""
-    if not key.startswith("_"):
-        return f"{domain}.{key}"
-    key = key[1:]  # Remove leading underscore
-    if not key:
-        return domain
-    return key
-
-
-def get_relative_description_key(domain: str, key: str) -> str:
-    """Return the relative description key."""
-    platform, *subtype = key.split(".", 1)
-    if platform != domain:
-        return f"_{key}"
-    if not subtype:
-        return "_"
-    return subtype[0]
-
-
 # Validator helpers
